requirements.txt
opencv-python
numpy
scipy
colorama
pillow
shapely